split up boot.lua like LÖVE does
