+++ include/modules/love.h
@@ -1,12 +1,16 @@
 #pragma once
 
-namespace Love
+namespace love
 {
     int Initialize(lua_State* L);
 
-    int Boot(lua_State* L);
-
     int NoGame(lua_State* L);
+
+    int LoadArgs(lua_State* L);
+
+    int LoadCallbacks(lua_State* L);
+
+    int Boot(lua_State* L);
 
     int GetVersion(lua_State* L);
 
@@ -14,10 +18,11 @@ namespace Love
 
     int IsVersionCompatible(lua_State* L);
 
-    /* init nxlink or something */
+    /* Debugging Utility */
+
     int _OpenConsole(lua_State* L);
 
-    /* -------- */
+    /* ----------------- */
 
     int Preload(lua_State* L, lua_CFunction func, const char* name);
-}; // namespace Love
+}; // namespace love
