+++ platform/3ds/Makefile
@@ -46,6 +46,7 @@ DATA		:=	${LOVE_DATA_FILES} \
 
 INCLUDES	?=	${LOVE_INCLUDES} \
 				${LOVE_LIBRARIES} \
+				${LOVE_MAIN_DATA_FILES} \
 				${CONSOLE_INCLUDE}
 
 GRAPHICS	:=	graphics
