+++ platform/switch/Makefile
@@ -53,6 +53,7 @@ DATA		:=	${LOVE_DATA_FILES} \
 
 INCLUDES	:=	${LOVE_INCLUDES} \
 				${LOVE_LIBRARIES} \
+				${LOVE_MAIN_DATA_FILES} \
 				${CONSOLE_INCLUDE}
 
 APP_TITLEID := 1043
